requirements.txt
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
Pillow==11.1.0
python-decouple==3.8
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.31.0 